+++ main.go
@@ -1,11 +1,15 @@
 package main
 
 import (
+	"errors"
 	"fmt"
+	"math/rand"
 	"os"
+	"strings"
 	"time"
 
 	"github.com/GannettDigital/go-newrelic-plugin/collectors"
+	"github.com/GannettDigital/goFigure"
 	"github.com/Sirupsen/logrus"
 	newrelicMonitoring "github.com/newrelic/go-agent"
 	"github.com/spf13/viper"
@@ -15,7 +19,10 @@ var log = logrus.New()
 
 func main() {
 
-	config := loadConfig()
+	config, err := loadConfig("", &gofigure.ConfigClient{}, os.Getenv("GOFIGURE_BUCKET"), os.Getenv("GOFIGURE_ITEM_PATH"))
+	if err != nil {
+		os.Exit(1)
+	}
 
 	app := setupNewRelic(config)
 
@@ -24,7 +31,10 @@ func main() {
 		// if config file has enabled the collector indicated by collectorName
 		if config.Collectors[name].Enabled {
 			go func(collectorName string, collectorValue collectors.Collector) {
-				// TODO: random delay to offset collections
+				rand.Seed(time.Now().UnixNano())
+				sleepTime := time.Millisecond * time.Duration(rand.Intn(1000))
+				log.Info(fmt.Sprintf("sleeping %s to offset collector collections", sleepTime))
+				time.Sleep(sleepTime)
 				ticker := time.NewTicker(readCollectorDelay(collectorName, config))
 				for _ = range ticker.C {
 					getResult(collectorName, app, config, collectorValue)
@@ -106,13 +116,49 @@ func sendData(collectorName string, app newrelicMonitoring.Application, config c
 	log.WithFields(logrus.Fields{
 		"collector": collectorName,
 	}).Info("recording event")
+
+	tags := processTags(collectorName, config)
+	payload := mergeMaps(tags, stats)
 	// send stats
-	app.RecordCustomEvent(fmt.Sprintf("gannettNewRelic%s", collectorName), stats)
+	app.RecordCustomEvent(fmt.Sprintf("gannettNewRelic%s", collectorName), payload)
+}
+
+// processTags - read all ENV tags, and append collector specific to global for both kv tags and env tags
+func processTags(collectorName string, config collectors.Config) map[string]interface{} {
+	kvList := mergeMaps(convertToInterfaceMap(config.Tags.KeyValue), convertToInterfaceMap(config.Collectors[collectorName].Tags.KeyValue))
+	envList := append(config.Tags.Env, config.Collectors[collectorName].Tags.Env...)
+	kvList = mergeMaps(kvList, readEnvList(envList))
+	return kvList
+}
+
+// mergeMaps - merge two map[string]interface{}
+func mergeMaps(global map[string]interface{}, specific map[string]interface{}) map[string]interface{} {
+	for key, value := range specific {
+		global[key] = value
+	}
+	return global
+}
+
+// readEnvList - read all environment variables and return them as a map[string]interface{}
+func readEnvList(envList []string) map[string]interface{} {
+	resultList := make(map[string]interface{})
+	for _, env := range envList {
+		resultList[strings.ToLower(env)] = os.Getenv(env)
+	}
+	return resultList
+}
+
+// convertToInterfaceMap - make map[string]string into a map[string]interface
+func convertToInterfaceMap(stringMap map[string]string) map[string]interface{} {
+	interfaceMap := make(map[string]interface{})
+	for key, value := range stringMap {
+		interfaceMap[key] = value
+	}
+	return interfaceMap
 }
 
 func setupNewRelic(config collectors.Config) newrelicMonitoring.Application {
 
-	// TODO: pull from config
 	// Create an app config.  Application name and New Relic license key are required.
 	cfg := newrelicMonitoring.NewConfig(config.AppName, config.NewRelicKey)
 
@@ -147,11 +193,17 @@ func setupNewRelic(config collectors.Config) newrelicMonitoring.Application {
 }
 
 // loadConfig - read from config file and marshal info collectors.Config
-func loadConfig() collectors.Config {
+func loadConfig(configName string, client gofigure.Client, bucket string, itemPath string) (collectors.Config, error) {
+	if configName == "" {
+		configName = "config"
+	}
+	// config object to return
+	var conf collectors.Config
+
 	// set up viper to find config file
 	vip := viper.New()
 	vip.SetConfigType("yaml")
-	vip.SetConfigName("config")
+	vip.SetConfigName(configName)
 	vip.AddConfigPath("/etc/newrelic_plugins/")
 	vip.AddConfigPath(".")
 
@@ -160,20 +212,36 @@ func loadConfig() collectors.Config {
 	if err != nil {
 		log.WithFields(logrus.Fields{
 			"err": err,
-		}).Error("Error reading config file")
+		}).Info("Error reading local config file, will attempt to locate goFigure configs...")
 
-		os.Exit(1)
+		// attempt to read from s3 bucket if GOFIGURE_BUCKET and GOFIGURE_ITEM_PATH are set
+		if (bucket != "") && (itemPath != "") {
+			client.Setup(bucket, itemPath, "yaml")
+
+			err = client.LoadAndUnmarshal(&conf)
+			if err != nil {
+				log.WithFields(logrus.Fields{
+					"err": err,
+				}).Error("Error loading config file from s3")
+
+				return conf, err
+			}
+
+			return conf, nil
+		}
+
+		log.WithFields(logrus.Fields{}).Error("Error locating any configs")
+		return conf, errors.New("No configs located")
 	}
 
 	// marshal config file data into collectors.Config and return it
-	var conf collectors.Config
 	err = vip.Unmarshal(&conf)
 	if err != nil {
 		log.WithFields(logrus.Fields{
 			"err": err,
 		}).Error("Error unmarshaling configs")
 
-		os.Exit(1)
+		return conf, err
 	}
-	return conf
+	return conf, nil
 }